refactor: Move OffsetStatus module to the offset-manager-api
- OffsetStatus is used for how to save in OffsetManager and also how to handle Offset in the source connector.
