+++ offset-manager-api/build.gradle
@@ -13,6 +13,7 @@ dependencies {
   testImplementation platform('org.junit:junit-bom:5.10.0')
   testImplementation 'org.junit.jupiter:junit-jupiter'
   testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+  compileOnly("org.projectlombok:lombok:1.18.38")
 }
 
 test {
